Move the conversion index list to utils and constants
