+++ python/symprobe/utils.py
@@ -13,6 +13,8 @@ import re
 import numpy as np
 import pandas as pd
 
+from symprobe.constants import CONVERSION_IDX
+
 
 def get_print_timestep(log_path):
     """Extracts the print timestep value from the log file
@@ -149,6 +151,26 @@ def get_range(num_range):
         num_list = [int(i) for i in num_range]
 
     return num_list
+
+
+def convert_connections(cube_node_list):
+    """Converts the connections of the cubic element to six tetrahedra
+    connections
+
+    Arguments:
+    cube_node_list -- list[int], list of nodes for the cubic element.
+
+    Return:
+    tet_node_list -- list[list[int]], list of the six node lists for the
+            tetrahedral elements.
+
+    """
+    tet_node_list = []
+    for idx_list in CONVERSION_IDX:
+        tet_list = [cube_node_list[idx] for idx in idx_list]
+        tet_node_list.append(tet_list)
+
+    return tet_node_list
 
 
 def print_quality(quality_array, metric_name):
